refactor: merge search from :app and :tv into common one
