--- app/src/main/java/tv/trakt/trakt/core/search/data/local/model/RecentMovieEntity.kt
@@ -1,23 +1,0 @@
-package tv.trakt.trakt.core.search.data.local.model
-
-import kotlinx.serialization.Serializable
-import tv.trakt.trakt.common.model.Movie
-import java.time.Instant
-
-@Serializable
-internal data class RecentMovieEntity(
-    val movie: Movie,
-    val createdAt: String, // Format: "2022-12-03T10:15:30Z"
-) {
-    companion object
-}
-
-internal fun RecentMovieEntity.Companion.create(
-    movie: Movie,
-    createdAt: Instant,
-): RecentMovieEntity {
-    return RecentMovieEntity(
-        movie = movie,
-        createdAt = createdAt.toString(),
-    )
-}
