+++ tv/src/main/java/tv/trakt/trakt/app/core/search/di/SearchModule.kt
@@ -2,38 +2,28 @@ package tv.trakt.trakt.app.core.search.di
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
-import androidx.lifecycle.SavedStateHandle
-import io.ktor.client.HttpClientConfig
-import io.ktor.client.engine.HttpClientEngine
 import org.koin.android.ext.koin.androidContext
-import org.koin.core.module.dsl.viewModel
+import org.koin.core.module.dsl.factoryOf
+import org.koin.core.module.dsl.viewModelOf
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
-import org.openapitools.client.apis.SearchApi
 import tv.trakt.trakt.app.core.main.di.createStore
 import tv.trakt.trakt.app.core.search.SearchViewModel
 import tv.trakt.trakt.app.core.search.data.local.RecentSearchLocalDataSource
 import tv.trakt.trakt.app.core.search.data.local.RecentSearchStorage
-import tv.trakt.trakt.app.core.search.data.remote.SearchApiClient
-import tv.trakt.trakt.app.core.search.data.remote.SearchRemoteDataSource
-import tv.trakt.trakt.app.core.search.usecase.GetSearchResultsUseCase
 import tv.trakt.trakt.app.core.search.usecase.recents.AddRecentSearchUseCase
 import tv.trakt.trakt.app.core.search.usecase.recents.GetRecentSearchUseCase
-import tv.trakt.trakt.common.Config.API_BASE_URL
+import tv.trakt.trakt.common.core.search.data.remote.SearchApiClient
+import tv.trakt.trakt.common.core.search.data.remote.SearchRemoteDataSource
+import tv.trakt.trakt.common.core.search.usecase.GetSearchResultsUseCase
 
 private const val SEARCH_PREFERENCES = "search_preferences_tv_v2"
 
 internal val searchDataModule = module {
     single<SearchRemoteDataSource> {
-        val httpClientEngine = get<HttpClientEngine>()
-        val httpClientConfig = get<(HttpClientConfig<*>) -> Unit>(named("clientConfig"))
-
         SearchApiClient(
-            api = SearchApi(
-                baseUrl = API_BASE_URL,
-                httpClientEngine = httpClientEngine,
-                httpClientConfig = httpClientConfig,
-            ),
+            api = get(),
+            authorizedApi = get(named("authorizedSearchApi")),
         )
     }
 
@@ -52,34 +42,9 @@ internal val searchDataModule = module {
 }
 
 internal val searchModule = module {
+    factoryOf(::GetSearchResultsUseCase)
+    factoryOf(::GetRecentSearchUseCase)
+    factoryOf(::AddRecentSearchUseCase)
 
-    factory {
-        GetSearchResultsUseCase(
-            remoteSource = get(),
-        )
-    }
-
-    factory {
-        AddRecentSearchUseCase(
-            recentsLocalSource = get(),
-        )
-    }
-
-    factory {
-        GetRecentSearchUseCase(
-            recentsLocalSource = get(),
-        )
-    }
-
-    viewModel { (_: SavedStateHandle) ->
-        SearchViewModel(
-            getSearchResultsUseCase = get(),
-            addRecentSearchUseCase = get(),
-            getRecentSearchUseCase = get(),
-            getTrendingShowsUseCase = get(),
-            getTrendingMoviesUseCase = get(),
-            showLocalSource = get(),
-            movieLocalSource = get(),
-        )
-    }
+    viewModelOf(::SearchViewModel)
 }
